requirements.txt
asgiref==3.7.2
Django==4.0.4
django-assets==2.0
django-ckeditor==6.5.1
django-colorfield==0.8.0
django-js-asset==2.1.0
django-localflavor==3.1
django-model-utils==4.3.1
djangorestframework==3.13.1
Pillow==10.0.0
python-stdnum==1.18
pytz==2023.3
sqlparse==0.4.4
typing_extensions==4.7.1
tzdata==2023.3
webassets==2.0
